refactor(types): decompose provider settings
