+++ packages/types/src/provider-settings.ts
@@ -1,12 +1,15 @@
 import { z } from "zod"
 
-import {
-	modelInfoSchema,
-	openAiCodexServiceTierSchema,
-	reasoningEffortSettingSchema,
-	verbosityLevelsSchema,
-	serviceTierSchema,
-} from "./model.js"
+import { providerDefinitionList, type ProviderDefinition } from "./provider-settings/index.js"
+import { API_PROVIDER_FIELD, SETTINGS_SHAPE_FIELD } from "./provider-settings/common.js"
+export {
+	OPEN_AI_CODEX_SERVICE_TIER_KEY,
+	kimiCodeAuthMethodSchema,
+	type KimiCodeAuthMethod,
+	zaiApiLineSchema,
+	type ZaiApiLine,
+} from "./provider-settings/index.js"
+
 import { codebaseIndexProviderSchema } from "./codebase-index.js"
 import type { UnionToIntersection } from "./type-fu.js"
 import {
@@ -45,7 +48,6 @@ import {
  */
 
 export const DEFAULT_CONSECUTIVE_MISTAKE_LIMIT = 3
-export const OPEN_AI_CODEX_SERVICE_TIER_KEY = "openAiCodexServiceTier"
 
 /**
  * DynamicProvider
@@ -161,20 +163,6 @@ export type ProviderNameWithRetired = z.infer<typeof providerNamesWithRetiredSch
  * ProviderSettingsEntry
  */
 
-const API_PROVIDER_FIELD = "apiProvider"
-const SETTINGS_SHAPE_FIELD = "settingsShape"
-const API_MODEL_ID_FIELD = "apiModelId"
-const OPEN_ROUTER_MODEL_ID_FIELD = "openRouterModelId"
-const OPEN_AI_MODEL_ID_FIELD = "openAiModelId"
-const OLLAMA_MODEL_ID_FIELD = "ollamaModelId"
-const LM_STUDIO_MODEL_ID_FIELD = "lmStudioModelId"
-const REQUESTY_MODEL_ID_FIELD = "requestyModelId"
-const UNBOUND_MODEL_ID_FIELD = "unboundModelId"
-const LITELLM_MODEL_ID_FIELD = "litellmModelId"
-const VERCEL_AI_GATEWAY_MODEL_ID_FIELD = "vercelAiGatewayModelId"
-const OPENCODE_GO_MODEL_ID_FIELD = "opencodeGoModelId"
-const ZOO_GATEWAY_MODEL_ID_FIELD = "zooGatewayModelId"
-
 export const providerSettingsEntrySchema = z.object({
 	id: z.string(),
 	name: z.string(),
@@ -187,551 +175,6 @@ export type ProviderSettingsEntry = z.infer<typeof providerSettingsEntrySchema>
 /**
  * ProviderSettings
  */
-
-const baseProviderSettingsShape = {
-	includeMaxTokens: z.boolean().optional(),
-	todoListEnabled: z.boolean().optional(),
-	modelTemperature: z.number().nullish(),
-	rateLimitSeconds: z.number().optional(),
-	consecutiveMistakeLimit: z.number().min(0).optional(),
-
-	// Model reasoning.
-	enableReasoningEffort: z.boolean().optional(),
-	reasoningEffort: reasoningEffortSettingSchema.optional(),
-	modelMaxTokens: z.number().optional(),
-	modelMaxThinkingTokens: z.number().optional(),
-
-	// Model verbosity.
-	verbosity: verbosityLevelsSchema.optional(),
-}
-
-type ModelId = string | undefined
-type UntypedProviderSettings = Record<string, unknown>
-type ProviderModelIdAccessor = (settings: UntypedProviderSettings) => ModelId
-
-type ProviderSettingsFromSchema<S extends z.ZodRawShape> = z.infer<z.ZodObject<S>>
-
-type TypedProviderModelIdAccessor<S extends z.ZodRawShape> = (settings: ProviderSettingsFromSchema<S>) => ModelId
-
-type ProviderDefinitionInput<P extends ProviderName, S extends z.ZodRawShape> = {
-	apiProvider: P
-	schema: S
-	modelIdKey?: Extract<keyof S, string>
-	getModelId: TypedProviderModelIdAccessor<S>
-}
-
-const createModelIdAccessor =
-	(modelIdKey: string): ProviderModelIdAccessor =>
-	(settings) =>
-		settings[modelIdKey] as ModelId
-
-// `modelIdKey` supports deprecated exports. Remove it in favor of an accessor-only contract when those exports are removed.
-const createProviderDefinition = <P extends ProviderName, S extends z.ZodRawShape>({
-	apiProvider,
-	schema,
-	...modelIdDefinition
-}: ProviderDefinitionInput<P, S>) => {
-	const settingsSchema = z.object(schema)
-	const getModelId: ProviderModelIdAccessor = (settings) => {
-		const parsedSettings = settingsSchema.safeParse(settings)
-		return parsedSettings.success ? modelIdDefinition.getModelId(parsedSettings.data) : undefined
-	}
-
-	return {
-		apiProvider,
-		settingsShape: schema,
-		modelIdKey: modelIdDefinition.modelIdKey,
-		schema: z.object({
-			...schema,
-			[API_PROVIDER_FIELD]: z.literal(apiProvider),
-		}),
-		getModelId,
-	}
-}
-
-// Several of the providers share common model config properties.
-const apiModelIdProviderModelShape = {
-	...baseProviderSettingsShape,
-	[API_MODEL_ID_FIELD]: z.string().optional(),
-}
-
-const anthropicProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.anthropic,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		apiKey: z.string().optional(),
-		anthropicBaseUrl: z.string().optional(),
-		anthropicUseAuthToken: z.boolean().optional(),
-		anthropicBeta1MContext: z.boolean().optional(), // Enable 'context-1m-2025-08-07' beta for 1M context window.
-	},
-})
-
-const openRouterProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.openrouter,
-	modelIdKey: OPEN_ROUTER_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(OPEN_ROUTER_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		openRouterApiKey: z.string().optional(),
-		[OPEN_ROUTER_MODEL_ID_FIELD]: z.string().optional(),
-		openRouterBaseUrl: z.string().optional(),
-		openRouterSpecificProvider: z.string().optional(),
-	},
-})
-
-const bedrockProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.bedrock,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		awsAccessKey: z.string().optional(),
-		awsSecretKey: z.string().optional(),
-		awsSessionToken: z.string().optional(),
-		awsRegion: z.string().optional(),
-		awsUseCrossRegionInference: z.boolean().optional(),
-		awsUseGlobalInference: z.boolean().optional(), // Enable Global Inference profile routing when supported
-		awsUsePromptCache: z.boolean().optional(),
-		awsProfile: z.string().optional(),
-		awsUseProfile: z.boolean().optional(),
-		awsApiKey: z.string().optional(),
-		awsUseApiKey: z.boolean().optional(),
-		awsCustomArn: z.string().optional(),
-		awsModelContextWindow: z.number().optional(),
-		awsBedrockEndpointEnabled: z.boolean().optional(),
-		awsBedrockEndpoint: z.string().optional(),
-		awsBedrock1MContext: z.boolean().optional(), // Enable 'context-1m-2025-08-07' beta for 1M context window.
-		awsBedrockServiceTier: z.enum(["STANDARD", "FLEX", "PRIORITY"]).optional(), // AWS Bedrock service tier selection
-	},
-})
-
-const vertexProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.vertex,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		vertexKeyFile: z.string().optional(),
-		vertexJsonCredentials: z.string().optional(),
-		vertexProjectId: z.string().optional(),
-		vertexRegion: z.string().optional(),
-		vertex1MContext: z.boolean().optional(), // Enable 'context-1m-2025-08-07' beta for 1M context window.
-	},
-})
-
-const openAiProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.openai,
-	modelIdKey: OPEN_AI_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(OPEN_AI_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		openAiBaseUrl: z.string().optional(),
-		openAiApiKey: z.string().optional(),
-		openAiR1FormatEnabled: z.boolean().optional(),
-		[OPEN_AI_MODEL_ID_FIELD]: z.string().optional(),
-		openAiCustomModelInfo: modelInfoSchema.nullish(),
-		openAiUseAzure: z.boolean().optional(),
-		azureApiVersion: z.string().optional(),
-		openAiStreamingEnabled: z.boolean().optional(),
-		openAiHostHeader: z.string().optional(), // Keep temporarily for backward compatibility during migration.
-		openAiHeaders: z.record(z.string(), z.string()).optional(),
-	},
-})
-
-const ollamaProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.ollama,
-	modelIdKey: OLLAMA_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(OLLAMA_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		[OLLAMA_MODEL_ID_FIELD]: z.string().optional(),
-		ollamaBaseUrl: z.string().optional(),
-		ollamaApiKey: z.string().optional(),
-		ollamaNumCtx: z.number().int().min(128).optional(),
-	},
-})
-
-const vsCodeLmProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.vscodeLm,
-	getModelId: (settings) => settings.vsCodeLmModelSelector?.id,
-	schema: {
-		...baseProviderSettingsShape,
-		vsCodeLmModelSelector: z
-			.object({
-				vendor: z.string().optional(),
-				family: z.string().optional(),
-				version: z.string().optional(),
-				id: z.string().optional(),
-			})
-			.optional(),
-	},
-})
-
-const lmStudioProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.lmstudio,
-	modelIdKey: LM_STUDIO_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(LM_STUDIO_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		[LM_STUDIO_MODEL_ID_FIELD]: z.string().optional(),
-		lmStudioBaseUrl: z.string().optional(),
-		lmStudioDraftModelId: z.string().optional(),
-		lmStudioSpeculativeDecodingEnabled: z.boolean().optional(),
-	},
-})
-
-const geminiProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.gemini,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		geminiApiKey: z.string().optional(),
-		googleGeminiBaseUrl: z.string().optional(),
-	},
-})
-
-const geminiCliProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.geminiCli,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		geminiCliOAuthPath: z.string().optional(),
-		geminiCliProjectId: z.string().optional(),
-	},
-})
-
-const openAiCodexProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.openaiCodex,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		// Codex "Fast" mode maps to the Responses API priority service tier.
-		[OPEN_AI_CODEX_SERVICE_TIER_KEY]: openAiCodexServiceTierSchema.optional(),
-	},
-})
-
-const openAiNativeProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.openaiNative,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		openAiNativeApiKey: z.string().optional(),
-		openAiNativeBaseUrl: z.string().optional(),
-		// OpenAI Responses API service tier for openai-native provider only.
-		// UI should only expose this when the selected model supports flex/priority.
-		openAiNativeServiceTier: serviceTierSchema.optional(),
-	},
-})
-
-const mistralProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.mistral,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		mistralApiKey: z.string().optional(),
-		mistralCodestralUrl: z.string().optional(),
-	},
-})
-
-const deepSeekProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.deepseek,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		deepSeekBaseUrl: z.string().optional(),
-		deepSeekApiKey: z.string().optional(),
-	},
-})
-
-const poeProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.poe,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		poeApiKey: z.string().optional(),
-		poeBaseUrl: z.string().optional(),
-	},
-})
-
-const moonshotProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.moonshot,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		moonshotBaseUrl: z
-			.union([z.literal("https://api.moonshot.ai/v1"), z.literal("https://api.moonshot.cn/v1")])
-			.optional(),
-		moonshotApiKey: z.string().optional(),
-	},
-})
-
-export const kimiCodeAuthMethodSchema = z.enum(["oauth", "api-key"])
-export type KimiCodeAuthMethod = z.infer<typeof kimiCodeAuthMethodSchema>
-
-const kimiCodeProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.kimiCode,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		kimiCodeAuthMethod: kimiCodeAuthMethodSchema.optional(),
-		kimiCodeApiKey: z.string().optional(),
-	},
-})
-
-const minimaxProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.minimax,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		minimaxBaseUrl: z
-			.union([z.literal("https://api.minimax.io/v1"), z.literal("https://api.minimaxi.com/v1")])
-			.optional(),
-		minimaxApiKey: z.string().optional(),
-	},
-})
-
-const mimoProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.mimo,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		mimoBaseUrl: z
-			.union([
-				z.literal("https://api.xiaomimimo.com/v1"),
-				z.literal("https://token-plan-cn.xiaomimimo.com/v1"),
-				z.literal("https://token-plan-sgp.xiaomimimo.com/v1"),
-				z.literal("https://token-plan-ams.xiaomimimo.com/v1"),
-			])
-			.optional(),
-		mimoApiKey: z.string().optional(),
-	},
-})
-
-const requestyProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.requesty,
-	modelIdKey: REQUESTY_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(REQUESTY_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		requestyBaseUrl: z.string().optional(),
-		requestyApiKey: z.string().optional(),
-		[REQUESTY_MODEL_ID_FIELD]: z.string().optional(),
-	},
-})
-
-const unboundProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.unbound,
-	modelIdKey: UNBOUND_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(UNBOUND_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		unboundApiKey: z.string().optional(),
-		[UNBOUND_MODEL_ID_FIELD]: z.string().optional(),
-	},
-})
-
-const fakeAiProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.fakeAi,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		fakeAi: z.unknown().optional(),
-	},
-})
-
-const xaiProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.xai,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		xaiApiKey: z.string().optional(),
-	},
-})
-
-const litellmProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.litellm,
-	modelIdKey: LITELLM_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(LITELLM_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		litellmBaseUrl: z.string().optional(),
-		litellmApiKey: z.string().optional(),
-		[LITELLM_MODEL_ID_FIELD]: z.string().optional(),
-		litellmUsePromptCache: z.boolean().optional(),
-	},
-})
-
-const sambaNovaProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.sambanova,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		sambaNovaApiKey: z.string().optional(),
-	},
-})
-
-export const zaiApiLineSchema = z.enum(["international_coding", "china_coding", "international_api", "china_api"])
-
-export type ZaiApiLine = z.infer<typeof zaiApiLineSchema>
-
-const zaiProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.zai,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		zaiApiKey: z.string().optional(),
-		zaiApiLine: zaiApiLineSchema.optional(),
-	},
-})
-
-const fireworksProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.fireworks,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		fireworksApiKey: z.string().optional(),
-	},
-})
-
-const friendliProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.friendli,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		friendliApiKey: z.string().optional(),
-	},
-})
-
-const qwenCodeProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.qwenCode,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		qwenCodeOauthPath: z.string().optional(),
-	},
-})
-
-const vercelAiGatewayProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.vercelAiGateway,
-	modelIdKey: VERCEL_AI_GATEWAY_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(VERCEL_AI_GATEWAY_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		vercelAiGatewayApiKey: z.string().optional(),
-		[VERCEL_AI_GATEWAY_MODEL_ID_FIELD]: z.string().optional(),
-	},
-})
-
-const opencodeGoProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.opencodeGo,
-	modelIdKey: OPENCODE_GO_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(OPENCODE_GO_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		opencodeGoApiKey: z.string().optional(),
-		[OPENCODE_GO_MODEL_ID_FIELD]: z.string().optional(),
-	},
-})
-
-const KENARI_MODEL_ID_FIELD = "kenariModelId"
-
-const kenariProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.kenari,
-	modelIdKey: KENARI_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(KENARI_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		kenariApiKey: z.string().optional(),
-		[KENARI_MODEL_ID_FIELD]: z.string().optional(),
-	},
-})
-
-const zooGatewayProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.zooGateway,
-	modelIdKey: ZOO_GATEWAY_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(ZOO_GATEWAY_MODEL_ID_FIELD),
-	schema: {
-		...baseProviderSettingsShape,
-		zooSessionToken: z.string().optional(),
-		[ZOO_GATEWAY_MODEL_ID_FIELD]: z.string().optional(),
-		zooGatewayBaseUrl: z.string().optional(),
-	},
-})
-
-const basetenProviderDefinition = createProviderDefinition({
-	apiProvider: providerIdentifiers.baseten,
-	modelIdKey: API_MODEL_ID_FIELD,
-	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
-	schema: {
-		...apiModelIdProviderModelShape,
-		basetenApiKey: z.string().optional(),
-	},
-})
-
-type ProviderDefinition = {
-	apiProvider: ProviderName
-	settingsShape: z.ZodRawShape
-	modelIdKey?: string
-	schema: z.ZodDiscriminatedUnionOption<typeof API_PROVIDER_FIELD>
-	getModelId: ProviderModelIdAccessor
-}
-
-const providerDefinitionList = [
-	anthropicProviderDefinition,
-	openRouterProviderDefinition,
-	bedrockProviderDefinition,
-	vertexProviderDefinition,
-	openAiProviderDefinition,
-	ollamaProviderDefinition,
-	vsCodeLmProviderDefinition,
-	lmStudioProviderDefinition,
-	geminiProviderDefinition,
-	geminiCliProviderDefinition,
-	openAiCodexProviderDefinition,
-	openAiNativeProviderDefinition,
-	mistralProviderDefinition,
-	deepSeekProviderDefinition,
-	poeProviderDefinition,
-	moonshotProviderDefinition,
-	kimiCodeProviderDefinition,
-	minimaxProviderDefinition,
-	mimoProviderDefinition,
-	requestyProviderDefinition,
-	unboundProviderDefinition,
-	fakeAiProviderDefinition,
-	xaiProviderDefinition,
-	basetenProviderDefinition,
-	litellmProviderDefinition,
-	sambaNovaProviderDefinition,
-	zaiProviderDefinition,
-	fireworksProviderDefinition,
-	friendliProviderDefinition,
-	qwenCodeProviderDefinition,
-	vercelAiGatewayProviderDefinition,
-	opencodeGoProviderDefinition,
-	kenariProviderDefinition,
-	zooGatewayProviderDefinition,
-] as const satisfies readonly ProviderDefinition[]
 
 type ListedProvider = (typeof providerDefinitionList)[number][typeof API_PROVIDER_FIELD]
 const allProvidersAreDefined: Exclude<ProviderName, ListedProvider> extends never ? true : never = true
