+++ packages/types/src/provider-settings/common.ts
@@ -1,0 +1,87 @@
+import { z } from "zod"
+
+import { reasoningEffortSettingSchema, verbosityLevelsSchema } from "../model.js"
+import type { ProviderIdentifier } from "../provider-identifiers.js"
+
+export const API_PROVIDER_FIELD = "apiProvider"
+export const SETTINGS_SHAPE_FIELD = "settingsShape"
+export const API_MODEL_ID_FIELD = "apiModelId"
+export const OPEN_ROUTER_MODEL_ID_FIELD = "openRouterModelId"
+export const OPEN_AI_MODEL_ID_FIELD = "openAiModelId"
+export const OLLAMA_MODEL_ID_FIELD = "ollamaModelId"
+export const LM_STUDIO_MODEL_ID_FIELD = "lmStudioModelId"
+export const REQUESTY_MODEL_ID_FIELD = "requestyModelId"
+export const UNBOUND_MODEL_ID_FIELD = "unboundModelId"
+export const LITELLM_MODEL_ID_FIELD = "litellmModelId"
+export const VERCEL_AI_GATEWAY_MODEL_ID_FIELD = "vercelAiGatewayModelId"
+export const OPENCODE_GO_MODEL_ID_FIELD = "opencodeGoModelId"
+export const ZOO_GATEWAY_MODEL_ID_FIELD = "zooGatewayModelId"
+export const OPEN_AI_CODEX_SERVICE_TIER_KEY = "openAiCodexServiceTier"
+
+export const baseProviderSettingsShape = {
+	includeMaxTokens: z.boolean().optional(),
+	todoListEnabled: z.boolean().optional(),
+	modelTemperature: z.number().nullish(),
+	rateLimitSeconds: z.number().optional(),
+	consecutiveMistakeLimit: z.number().min(0).optional(),
+	enableReasoningEffort: z.boolean().optional(),
+	reasoningEffort: reasoningEffortSettingSchema.optional(),
+	modelMaxTokens: z.number().optional(),
+	modelMaxThinkingTokens: z.number().optional(),
+	verbosity: verbosityLevelsSchema.optional(),
+}
+
+export const apiModelIdProviderModelShape = {
+	...baseProviderSettingsShape,
+	[API_MODEL_ID_FIELD]: z.string().optional(),
+}
+
+type ModelId = string | undefined
+type UntypedProviderSettings = Record<string, unknown>
+type ProviderModelIdAccessor = (settings: UntypedProviderSettings) => ModelId
+type ProviderSettingsFromSchema<S extends z.ZodRawShape> = z.infer<z.ZodObject<S>>
+type TypedProviderModelIdAccessor<S extends z.ZodRawShape> = (settings: ProviderSettingsFromSchema<S>) => ModelId
+
+type ProviderDefinitionInput<P extends ProviderIdentifier, S extends z.ZodRawShape> = {
+	apiProvider: P
+	schema: S
+	modelIdKey?: Extract<keyof S, string>
+	getModelId: TypedProviderModelIdAccessor<S>
+}
+
+export type ProviderDefinition = {
+	apiProvider: ProviderIdentifier
+	settingsShape: z.ZodRawShape
+	modelIdKey?: string
+	schema: z.ZodDiscriminatedUnionOption<typeof API_PROVIDER_FIELD>
+	getModelId: ProviderModelIdAccessor
+}
+
+export const createModelIdAccessor =
+	(modelIdKey: string): ProviderModelIdAccessor =>
+	(settings) =>
+		settings[modelIdKey] as ModelId
+
+// `modelIdKey` supports deprecated exports. Remove it in favor of an accessor-only contract when those exports are removed.
+export const createProviderDefinition = <P extends ProviderIdentifier, S extends z.ZodRawShape>({
+	apiProvider,
+	schema,
+	...modelIdDefinition
+}: ProviderDefinitionInput<P, S>) => {
+	const settingsSchema = z.object(schema)
+	const getModelId: ProviderModelIdAccessor = (settings) => {
+		const parsedSettings = settingsSchema.safeParse(settings)
+		return parsedSettings.success ? modelIdDefinition.getModelId(parsedSettings.data) : undefined
+	}
+
+	return {
+		apiProvider,
+		settingsShape: schema,
+		modelIdKey: modelIdDefinition.modelIdKey,
+		schema: z.object({
+			...schema,
+			[API_PROVIDER_FIELD]: z.literal(apiProvider),
+		}),
+		getModelId,
+	}
+}
