+++ packages/types/src/provider-settings/friendli.ts
@@ -1,0 +1,19 @@
+import { z } from "zod"
+
+import { providerIdentifiers } from "../provider-identifiers.js"
+import {
+	API_MODEL_ID_FIELD,
+	apiModelIdProviderModelShape,
+	createModelIdAccessor,
+	createProviderDefinition,
+} from "./common.js"
+
+export const friendliProviderDefinition = createProviderDefinition({
+	apiProvider: providerIdentifiers.friendli,
+	modelIdKey: API_MODEL_ID_FIELD,
+	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
+	schema: {
+		...apiModelIdProviderModelShape,
+		friendliApiKey: z.string().optional(),
+	},
+})
