+++ packages/types/src/provider-settings/ollama.ts
@@ -1,0 +1,22 @@
+import { z } from "zod"
+
+import { providerIdentifiers } from "../provider-identifiers.js"
+import {
+	OLLAMA_MODEL_ID_FIELD,
+	baseProviderSettingsShape,
+	createModelIdAccessor,
+	createProviderDefinition,
+} from "./common.js"
+
+export const ollamaProviderDefinition = createProviderDefinition({
+	apiProvider: providerIdentifiers.ollama,
+	modelIdKey: OLLAMA_MODEL_ID_FIELD,
+	getModelId: createModelIdAccessor(OLLAMA_MODEL_ID_FIELD),
+	schema: {
+		...baseProviderSettingsShape,
+		[OLLAMA_MODEL_ID_FIELD]: z.string().optional(),
+		ollamaBaseUrl: z.string().optional(),
+		ollamaApiKey: z.string().optional(),
+		ollamaNumCtx: z.number().int().min(128).optional(),
+	},
+})
