+++ packages/types/src/provider-settings/openai-codex.ts
@@ -1,0 +1,20 @@
+import { providerIdentifiers } from "../provider-identifiers.js"
+import { openAiCodexServiceTierSchema } from "../model.js"
+import {
+	API_MODEL_ID_FIELD,
+	OPEN_AI_CODEX_SERVICE_TIER_KEY,
+	apiModelIdProviderModelShape,
+	createModelIdAccessor,
+	createProviderDefinition,
+} from "./common.js"
+
+export const openAiCodexProviderDefinition = createProviderDefinition({
+	apiProvider: providerIdentifiers.openaiCodex,
+	modelIdKey: API_MODEL_ID_FIELD,
+	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
+	schema: {
+		...apiModelIdProviderModelShape,
+		// Codex "Fast" mode maps to the Responses API priority service tier.
+		[OPEN_AI_CODEX_SERVICE_TIER_KEY]: openAiCodexServiceTierSchema.optional(),
+	},
+})
