+++ packages/types/src/provider-settings/openrouter.ts
@@ -1,0 +1,22 @@
+import { z } from "zod"
+
+import { providerIdentifiers } from "../provider-identifiers.js"
+import {
+	OPEN_ROUTER_MODEL_ID_FIELD,
+	baseProviderSettingsShape,
+	createModelIdAccessor,
+	createProviderDefinition,
+} from "./common.js"
+
+export const openRouterProviderDefinition = createProviderDefinition({
+	apiProvider: providerIdentifiers.openrouter,
+	modelIdKey: OPEN_ROUTER_MODEL_ID_FIELD,
+	getModelId: createModelIdAccessor(OPEN_ROUTER_MODEL_ID_FIELD),
+	schema: {
+		...baseProviderSettingsShape,
+		openRouterApiKey: z.string().optional(),
+		[OPEN_ROUTER_MODEL_ID_FIELD]: z.string().optional(),
+		openRouterBaseUrl: z.string().optional(),
+		openRouterSpecificProvider: z.string().optional(),
+	},
+})
