+++ packages/types/src/provider-settings/poe.ts
@@ -1,0 +1,20 @@
+import { z } from "zod"
+
+import { providerIdentifiers } from "../provider-identifiers.js"
+import {
+	API_MODEL_ID_FIELD,
+	apiModelIdProviderModelShape,
+	createModelIdAccessor,
+	createProviderDefinition,
+} from "./common.js"
+
+export const poeProviderDefinition = createProviderDefinition({
+	apiProvider: providerIdentifiers.poe,
+	modelIdKey: API_MODEL_ID_FIELD,
+	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
+	schema: {
+		...apiModelIdProviderModelShape,
+		poeApiKey: z.string().optional(),
+		poeBaseUrl: z.string().optional(),
+	},
+})
