+++ packages/types/src/provider-settings/sambanova.ts
@@ -1,0 +1,23 @@
+import { z } from "zod"
+
+import { providerIdentifiers } from "../provider-identifiers.js"
+import {
+	API_MODEL_ID_FIELD,
+	apiModelIdProviderModelShape,
+	createModelIdAccessor,
+	createProviderDefinition,
+} from "./common.js"
+
+export const sambaNovaProviderDefinition = createProviderDefinition({
+	apiProvider: providerIdentifiers.sambanova,
+	modelIdKey: API_MODEL_ID_FIELD,
+	getModelId: createModelIdAccessor(API_MODEL_ID_FIELD),
+	schema: {
+		...apiModelIdProviderModelShape,
+		sambaNovaApiKey: z.string().optional(),
+	},
+})
+
+export const zaiApiLineSchema = z.enum(["international_coding", "china_coding", "international_api", "china_api"])
+
+export type ZaiApiLine = z.infer<typeof zaiApiLineSchema>
